Moved / Pretty-ifing
Moved the sound timers to a function in MDTConfig called prepSounds.
Moved the private variables to MDTConfig with prepSounds.
Added whitespace to make the code more readable.
